requirements.txt
asgiref==3.4.1
Django==3.2.7
djangorestframework==3.12.4
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
